Minor: Split `DmlStatement` into its own module
